-Se cambia de int a long el numero de documento de nota credito y debito -Se organiza conversion de datos de ubl a objeto en el numero de documento.
git-svn-id: http://192.168.0.251:18080/svn/HGInetMiFacturaElectronica/Tronco@673 30a10932-8c21-4e4d-906a-15261f94da71

+++ HGInetMiFacturaElectonicaData/ModeloServicio/Documentos/NotaCredito.cs
@@ -31,7 +31,7 @@ namespace HGInetMiFacturaElectonicaData.ModeloServicio
         /// Numero de Documento
         /// </summary>
         [Required(AllowEmptyStrings = false, ErrorMessage = "El campo {0} es obligatorio")]
-        public int Documento { get; set; }
+        public long Documento { get; set; }
         
         /// <summary>
         /// Documento afectado
